refactor(algorithms): split algorithms directory in to "algorithms" and "dictionaries"
