+++ src/dictionaries/Algorithm.js
