+++ src/dictionaries/KeyAlgorithm.js
