+++ src/dictionaries/ShaKeyAlgorithm.js
