+++ test/algorithms/supportedAlgorithmsSpec.js
@@ -14,8 +14,8 @@ chai.should()
 const supportedAlgorithms = require('../../src/algorithms/supportedAlgorithms')
 const SupportedAlgorithms = supportedAlgorithms.constructor
 const RegisteredAlgorithms = require('../../src/algorithms/RegisteredAlgorithms')
-const RsaHashedKeyAlgorithm = require('../../src/algorithms/RsaHashedKeyAlgorithm')
-const ShaKeyAlgorithm = require('../../src/algorithms/ShaKeyAlgorithm')
+const RsaHashedKeyAlgorithm = require('../../src/dictionaries/RsaHashedKeyAlgorithm')
+const ShaKeyAlgorithm = require('../../src/dictionaries/ShaKeyAlgorithm')
 const NotSupportedError = require('../../src/errors/NotSupportedError')
 
 /**
